feat: reuse server types in ui
