+++ Cargo.toml
@@ -45,4 +45,5 @@ wasm-bindgen = "0.2"
 wasm-bindgen-test = "0.3"
 web-sys = "0.3"
 
+ipchat = { path = "src/server", default-features = false }
 ipchat-client = { path = "src/client" }
