+++ src/ui/src/hooks/chat_websocket.rs
@@ -1,6 +1,9 @@
 use leptos::prelude::{Get, Signal, expect_context};
+use uuid::Uuid;
 
-use crate::contexts::chat_websocket_context::{ChatWebSocketContext, Message, Peer, Room};
+use ipchat::proto::{ChatMessage, PeerRoom, Room};
+
+use crate::contexts::chat_websocket_context::ChatWebSocketContext;
 use crate::services::chat_websocket_service::ConnectionStatus;
 
 pub fn use_chat_ws() -> ChatWebSocketContext {
@@ -19,7 +22,7 @@ pub fn use_my_rooms() -> Signal<Vec<Room>> {
     Signal::derive(move || use_chat_ws().my_rooms.get())
 }
 
-pub fn use_discovered_peers() -> Signal<Vec<Peer>> {
+pub fn use_discovered_peers() -> Signal<Vec<PeerRoom>> {
     Signal::derive(move || use_chat_ws().discovered_peers.get())
 }
 
@@ -27,6 +30,6 @@ pub fn use_active_room() -> Signal<Option<Room>> {
     Signal::derive(move || use_chat_ws().active_room.get())
 }
 
-pub fn use_messages() -> Signal<std::collections::HashMap<String, Vec<Message>>> {
+pub fn use_messages() -> Signal<std::collections::HashMap<Uuid, Vec<ChatMessage>>> {
     Signal::derive(move || use_chat_ws().messages.get())
 }
